requirements.txt
faiss_gpu==1.7.2
gradio==5.15.0
langchain_huggingface==0.1.2
matplotlib==3.10.0
numpy==2.2.2
openai==1.61.1
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
youtube_transcript_api==0.6.3
